backend/requirements/all.txt
# ![IMPORTANT]: Only use if common.txt or requirements.txt doesn't work.
# ALL DEPENDICIES INCLUDING SUB PACKAGES  

anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
constantly==15.1.0
coverage==5.5
cryptography==41.0.4
cssbeautifier==1.14.9
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.6
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
EditorConfig==0.12.3
executing==2.0.1
Faker==19.10.0
fastjsonschema==2.19.0
filetype==1.2.0
fqdn==1.5.1
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jsbeautifier==1.14.9
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.6.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack==1.0.7
mutagen==1.47.0
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg==3.1.16
psycopg-binary==3.1.16
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
yarg==0.1.9
zope.interface==6.1

